backend/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
ortools==9.8.3296
tabulate==0.9.0